pyproject.toml
[project]
name = "sentiment-production-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "boto3>=1.42.29",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.8.1",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "transformers>=4.57.5",
    "uvicorn[standard]>=0.40.0",
    "pyyaml>=6.0.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
